Pull logging into its own file
